[NCL-8138] Add support for service account for callback
The service account is used to obtain an access token from the OIDC
server. This token is used for the callback that pnc-build-agent will
submit to the caller on completion.

In the past, we could rely on the caller to setup the access token to
use on the callback. However, we are going to reduce our access token
duration from days to 5 minutes, which means that we cannot use this
method anymore.

Instead, we need pnc-build-agent to generate a fresh access token when
sending the callback data to the caller.

If the `-keycloakClientConfig` CLI flag is not set, the access token
header won't be added in the callback request.